SamplePay: Create a sample payment app
This is a sample payment app.

Also add an empty payment web service that is to be associated with the Android app. The service is
not fully implemented and will not work on desktop, but it does not matter on mobile as far as the
mobile app is associated with it in manifest.json.

Change-Id: I73cceb066b91eadcc4384bc8362a0895c0a2963e